feat: add curriculum tracking events to store
